Mms:Fix Message content display blank at folder mode
When setting the pendingIntentTemplate for remote view with a sepecific
intent action(ACTION_VIEW) and type(vnd.android-dir/mms-sms),
the onClickFillInIntent set on each list item will not work.

In order to handle the click event on each list item(onClickFillInIntent),
do not specify any action or type when setting the pendingIntentTemplate
for the remote view.
Also, since the setOnClickPendingIntent() will not work on list view,
(This is a bug we met in CR-531038)
We change it to setOnClickFillInIntent() instead when setting up the list
item which is not loaded from the database yet(need display
"view more conversations").

Change-Id: I5858ebcf31464f8b51f67132203ee412d9455c7e
CRs-Fixed: 697232