Mms: Fix for adding too many recipients to the Message
There is a lot of jank when the user adds a large number of
recipients to a message. This is because of the appends required
to notify the RecipientsEditTextView of the new contacts that
need to be "chipped." This is results in an unmanageable amount of
work being done on the main thread. This workaround turns off
chips when the user adds 15 contacts. RecipientsEditTextView already
has an internal max of 50, but this is too high for this batched
recipient entry.

Change-Id: I1a292302ec783ebf75f48ab4aeac30645297df1f