Fixed KeyGuard not being disabled on second quickmessage
The keyguard was being properly hidden on the first time quickmessage
was being opened, however, after the application was opened that first
time, it would no longer disable the keyguard. There were some
unnecessary check to the keyguard and checking if the screen was on
which were causing the problem.

Change-Id: Ib13b253c6655f0955ed77018ff0a6ee67ea52501
Ticket: QRDL-940