Drafts: Fix instance where body of the message is lost.
WorkingMessage.loadDraft() can return a new WorkingMessage object that does not
have the logical mText. This happens when we have gone to select contacts to send
A message to and have loaded the previous draft upon our return in instances where the
user filled out a text message body before filling out any recipients.

Also, prevent the unnecessary settext when the length of the text is 0.

Change-Id: Ifc90f496476025058daf5697e4753a9f3707173d
Ticket: QRDL-957